Extract Batch 2 tool handlers into src/tools/
Extract detect_project, get_project_identity, set_project_identity,
list_migrations, execute_migration, get_project_memory_policy, and
set_project_memory_policy from index.ts into dedicated module files.

Batch 2 of index.ts modularization: Simple tool handlers.
